requirements.txt
aiobotocore==2.4.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
anyjson==0.3.3
apache-airflow==2.3.4
apache-airflow-providers-amazon==5.1.0
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-postgres==5.2.1
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
awswrangler==0.3.1
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.11.1
blinker==1.4
boto3==1.24.59
botocore==1.27.59
brotlipy==0.7.0
cachelib==0.9.0
cattrs==22.1.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
clickclick==1.2.2
colorama==0.4.5
colorlog==4.8.0
colour==0.1.5
commonmark==0.9.1
connexion==2.14.1
cron-descriptor==1.2.24
croniter==1.3.7
cryptography==37.0.4
dataclasses==0.8
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.4
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
exceptiongroup==1.0.0rc9
filelock==3.7.1
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flit_core==3.7.1
frozenlist==1.3.1
fsspec==2022.8.2
furl==2.1.3
future==0.18.2
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.9.0
infinity==1.5
inflection==0.5.1
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonschema==4.16.0
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
multidict==6.0.2
mypy-boto3-appflow==1.24.36.post1
mypy-boto3-rds==1.24.61
mypy-boto3-redshift-data==1.24.36.post1
numpy==1.23.2
orderedmultidict==1.0.1
packaging==21.3
pandas==1.4.4
passlib==1.7.4
pathspec==0.9.0
pendulum==2.1.2
pg8000==0.0.0
phonenumbers==8.12.55
pip==22.2.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
prison==0.2.1
psutil==5.9.2
psycopg2==2.9.3
pyarrow==9.0.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.908
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
s3fs==2022.8.2
s3transfer==0.6.0
scramp==1.4.1
setproctitle==1.3.2
setuptools==65.3.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.11
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
swagger-ui-bundle==0.0.6
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
typing_extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
Unidecode==1.3.4
urllib3==1.26.11
watchtower==2.0.1
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
WTForms==2.3.3
yarl==1.7.2
zipp==3.8.1
